refactor: Create matching utils
